base-requirements.txt
amqp==1.4.9
anyjson==0.3.3
astroid==1.4.9
bandit==1.1.0
billiard==3.3.0.23
block-disposable-email==1.0.1
celery==3.1.25
cffi==1.9.1
colorama==0.3.7
contextlib2==0.5.4
cryptography==1.7.2
Cython==0.25.2
Django==1.8.15
django-celery==3.2.1
django-disposable-email-checker==1.2.1
django-extensions==1.7.4
django-ipware==1.1.6
django-rest-knox==2.2.0
git+https://github.com/Seedstars/django-rest-logger.git
django-security==0.9.3
djangorestframework==3.4.7
docker==2.1.0
docker-pycreds==0.2.1
docopt==0.6.2
dodgy==0.1.9
gitdb2==2.0.0
GitPython==2.1.3
idna==2.5
isort==4.2.5
jedi==0.10.0
kombo==3.0.37
lazy-object-proxy==1.2.2
mccabe==0.6.1
packaging==16.8
pbr==2.0.0
pep8==1.7.0
pep8-naming==0.4.1
prompt-toolkit==1.0.13
prospector==0.12.3
psycopg2==2.6.2
ptpython==0.35
pyasn1==0.2.3
pycodestyle==2.0.0
pycparser==2.17
pydocstyle==1.0.0
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.6.5
pylint-celery==0.3
pylint-common==0.2.2
pylint-django==0.7.2
pylint-flask==0.5
pylint-plugin-utils==0.2.4
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-dateutil==2.5.3
PyYAML==3.12
raven==5.27.1
requests==2.13.0
requirements-detector==0.5.2
setoptconf==0.2.0
six==1.10.0
smmap2==2.0.1
stevedore==1.21.0
wcwidth==0.1.7
websocket-client==0.40.0
whitenoise==3.2.2
wrapt==1.10.8
